There are a few nice things though...
People criticized the NVIDIA API because they kind of brute-forced
their extensions, however I have to admit that for selecting devices,
this method is not THAT bad.

I mean, if you compare all the DRM/KMS/OpenGL code that goes like
"Here's our nice way to probe for devices" and then try to probe
EVERY SINGLE GPU DEVICE DRIVER MODULE NAME under the sun, in order
to get the DRM device filepath associated to it, the EGL way proposed
by NVIDIA isn't THAT BAD.

Now, I provided some code to the glmark2 codebase that mimicked Kwin
way for device probing. The whole thing used udev to get the properties
of DRM devices present, and tried to chose the right one based on
multiple criterias and... there's a lot of bug reports that want to
go BACK to the old "probe every fucking module name" way because they
have strangely weird setup that makes it hard to differentiate the
desired GPU from the others.

That said, I'm not talking about EGLStreams here... I still don't see
what *that* brings on the table.

Signed-off-by: Myy Miouyouyou <myy@miouyouyou.fr>