ALMOST THERE !
It's just that glClearColor is not working...
But glGetError() returns 0.

Is it supposed to work anyway ?

Is this shit done correctly NVIDIA ?

Between undocumented flags and functions returned values that makes
no sense compared to the documentation, it's hard to understand
what's going on...

Maybe I could use apitrace there and see what's going on.

Signed-off-by: Myy Miouyouyou <myy@miouyouyou.fr>