Ok, I'm done with this for the moment.
I still haven't tested alpha tests in OpenGL ES, but with a context
providing 0 alpha buffer, I really doubt that could work.
Of course, you can always generate draw buffers manually, ensure that
they have an alpha and depth buffer associated, draw in them and then
compose the screen with the buffers.

The whole idea of that kind of API is to exactly do this : Composition.

Still, there's overlay planes with ALPHA blending available, however
I have no idea about how to exploit them, and exploiting them with a
closed-source driver just drives me mad.

So I'll just remove the cruft in a few days and post it as a GIST.

Signed-off-by: Myy Miouyouyou <myy@miouyouyou.fr>